상품가격 소수점 가능 & 상위대리점 지정 가능

+++ src/main/java/com/gms/web/admin/controller/manage/ProductController.java
@@ -103,10 +103,10 @@ public class ProductController {
 				
 				//priceVo1.setProductId(Integer.valueOf(productId));
 				priceVo1.setProductCapa(req.getParameter("productCapa_"+i));
-				priceVo1.setProductPrice(Integer.parseInt(req.getParameter("productPrice_"+i)));				
+				priceVo1.setProductPrice(Float.parseFloat(req.getParameter("productPrice_"+i)));				
 				priceVo1.setECountCd(req.getParameter("eCountCd_"+i));
 				if(req.getParameter("productBottlePrice_"+i)!=null && req.getParameter("productBottlePrice_"+i).length() > 0)
-					priceVo1.setProductBottlePrice(Integer.parseInt(req.getParameter("productBottlePrice_"+i)));	
+					priceVo1.setProductBottlePrice(Float.parseFloat(req.getParameter("productBottlePrice_"+i)));	
 				else
 					priceVo1.setProductBottlePrice(0);
 				priceVo1.setECountCdS(req.getParameter("eCountCdS_"+i));
@@ -180,7 +180,7 @@ public class ProductController {
 			
 			RequestUtils.initUserPrgmInfo(req, params);
 			
-			boolean result=true;
+			boolean result=true;		
 			
 			params.setProductId(Integer.parseInt(req.getParameter("productId")));	
 			params.setProductNm(req.getParameter("productNm"));			
@@ -211,10 +211,10 @@ public class ProductController {
 					priceVo1.setProductId(Integer.parseInt(req.getParameter("productId")));					
 					priceVo1.setProductPriceSeq(Integer.parseInt(req.getParameter("productPriceSeq_"+i)));		
 					priceVo1.setProductCapa(req.getParameter("productCapa_"+i));
-					priceVo1.setProductPrice(Integer.parseInt(req.getParameter("productPrice_"+i)));					
+					priceVo1.setProductPrice(Float.parseFloat(req.getParameter("productPrice_"+i)));					
 					priceVo1.setECountCd(req.getParameter("eCountCd_"+i));
 					if(req.getParameter("productBottlePrice_"+i)!=null && req.getParameter("productBottlePrice_"+i).length() > 0)
-						priceVo1.setProductBottlePrice(Integer.parseInt(req.getParameter("productBottlePrice_"+i)));		
+						priceVo1.setProductBottlePrice(Float.parseFloat(req.getParameter("productBottlePrice_"+i)));		
 					else
 						priceVo1.setProductBottlePrice(0);	
 					priceVo1.setECountCdS(req.getParameter("eCountCdS_"+i));
@@ -227,9 +227,9 @@ public class ProductController {
 						priceVo1.setProductId(Integer.parseInt(req.getParameter("productId")));			
 						priceVo1.setProductPriceSeq(++lastPriceSeq);	
 						priceVo1.setProductCapa(req.getParameter("productCapa_"+i));
-						priceVo1.setProductPrice(Integer.parseInt(req.getParameter("productPrice_"+i)));						
+						priceVo1.setProductPrice(Float.parseFloat(req.getParameter("productPrice_"+i)));						
 						priceVo1.setECountCd(req.getParameter("eCountCd"+i));
-						priceVo1.setProductBottlePrice(Integer.parseInt(req.getParameter("productBottlePrice_"+i)));						
+						priceVo1.setProductBottlePrice(Float.parseFloat(req.getParameter("productBottlePrice_"+i)));						
 						priceVo1.setECountCdS(req.getParameter("eCountCdS"+i));
 						priceVo1.setCreateId(params.getCreateId());
